# Improve invalidation/validation performance