Initial:

|     Method |      Mean |     Error |    StdDev |
|----------- |----------:|----------:|----------:|
| Invalidate | 2.4286 ns | 0.0092 ns | 0.0072 ns |
|   Validate | 0.8363 ns | 0.0022 ns | 0.0020 ns |

Removing exception:

|     Method |      Mean |     Error |    StdDev |
|----------- |----------:|----------:|----------:|
| Invalidate | 0.5200 ns | 0.0233 ns | 0.0218 ns |
|   Validate | 0.8370 ns | 0.0063 ns | 0.0056 ns |

Removing branching:

|     Method |      Mean |     Error |    StdDev |
|----------- |----------:|----------:|----------:|
| Invalidate | 0.5070 ns | 0.0161 ns | 0.0143 ns |
|   Validate | 0.5107 ns | 0.0126 ns | 0.0098 ns |

I suspect the branching is different on different CPUs and different workloads, but will now be consistent across the board.

I expect this can be further vectorised as future work, but probably won't lead to much improvement in this case.