@smoogipoo 2022-09-06
Just randomly profiled one of my storyboard test beatmaps and `Drawable.invalidate()` (i.e. the whole thing) being 6% of total time.

Before:
![image](https://user-images.githubusercontent.com/1329837/188643524-9aa00dcf-5120-46d1-83c4-fb5a8a0bc4d9.png)

After:
![image](https://user-images.githubusercontent.com/1329837/188643566-47a3fe07-9554-4c51-bda7-2eba690e6edd.png)